Implement moment translation to CG and updated structure format
Looking at BAL_forces it can be seen that the X,Y,Z moments and forces
are the forces computed in some axis system (not sure where).
The moments are later translated to the moment point in the balance
axis system [0, 0, 0.0465/c]. Afterwards, these moments in still in X,Y,Z
are expressed in the airplane axis system (corrected for AoA). As such,
the location of the moments are already located in [0.25c, 0, 0] of the
aircraft reference frame as seen in the computatio nf o CMp25c.
It was desired to relocate these at the CG [0.48c, 0, 0]. Therefore,
it was only necessary to translate the pitching moment as the moment
arms for the yawing and rolling moments remain the same.

Included tau2 and delta correction factors inside load2Struct and
included additional parameters of the aircraft in the structure.