Add upwash lift interference correction function
Computes correction on inteference due to upwash
and upwash gradient or what Barlow, et. al. call
streamline curvature corrections.

It is pretty slow since I (wanted) to distinguish
the lift slope for different sideslip and free-
stream velocity.
If instead we were to assume that sideslip effect
is negligible, then there would be a maximum
error of 3%.