Add Eckert method to obtain thrust-free CL and CD, remove commented code
Implement Eckert's method to also obtain the thrust-free CL and CD.
Previously, the Eckert method was used to determine the thrust which
includes the induced drag. To subsequently use the lift interference and
wake blockage corrections, one cannot use the CL and CD which have the
thrust effects as it violates one of the assumptions in these
corrections. As such, thrust-free equivalent values must be obtained.

Note that slipstream corrections do not require that the thrust-free
should be computed.

Additional tail properties are added in load2Struct.

main.m should be nearly finished. A TODO is to remove the model off data
which Thomas should be working on / implementing.