Remove model off effects on measurements
Near-exact copy of removeModelOff method in the removeThrustUpdate
class. But the angle-of-attack can now be varied. Did not touch
the original method to prevent any bugs from happening. Note that
interp1 should be used when not a predictable / expected trend is
expected. Since polyfit is used with order 9, one order larger than
the total number of unique measurement points, this should be fine.
However, for more robust method, use interp1.